Stripped down starter files
these are files from a wordpress site I made that I have stripped down
so that we can edit them. These are files that wordpress needs in order
to make the theme. when we finish with the code, we will zip them up and
upload them to our wordpress instances. If you have any questions,call
me. You all have my cell number.